requirements.txt
django==3.2.16
djangorestframework==3.14.0
datetime==4.7
psycopg2==2.9.5
psycopg2-binary==2.9.5